[cleanup] Move events_table_list_view.h into wxcrafter/UI/
